add DB test

+++ db/migrate/20190729081652_create_locations.rb
@@ -2,8 +2,8 @@ class CreateLocations < ActiveRecord::Migration[5.2]
   def change
     create_table :locations do |t|
       t.string :name
-      t.float :latitude
-      t.float :longitude
+      t.numeric :latitude
+      t.numeric :longitude
 
       t.timestamps
     end
